+++ packages/contact-center/cc-components/webpack.config.js
@@ -1,4 +1,4 @@
-const { mergeWithCustomize, customizeObject } = require('webpack-merge');
+const {mergeWithCustomize, customizeObject} = require('webpack-merge');
 const path = require('path');
 
 const baseConfig = require('../../../webpack.config');
@@ -8,8 +8,8 @@ const resolveMonorepoRoot = (...segments) => path.resolve(__dirname, '../../../'
 
 module.exports = mergeWithCustomize({
   customizeObject: customizeObject({
-    'resolve.fallback': 'replace' // This will replace the fallback configuration
-  })
+    'resolve.fallback': 'replace', // This will replace the fallback configuration
+  }),
 })(baseConfig, {
   entry: {
     wc: {
@@ -29,10 +29,10 @@ module.exports = mergeWithCustomize({
     react: 'react',
     'react-dom': 'react-dom',
     '@webex/cc-store': '@webex/cc-store',
-    '@momentum-design/components': '@momentum-design/components',
+    '@momentum-ui/react-collaboration': '@momentum-ui/react-collaboration',
   },
   resolve: {
-    fallback: {}
+    fallback: {},
   },
   module: {
     rules: [
@@ -40,38 +40,39 @@ module.exports = mergeWithCustomize({
         test: /\.css$/,
         use: ['style-loader', 'css-loader'],
         include: [
-          resolveMonorepoRoot('node_modules/@momentum-design'), // Include specific node module
-          path.resolve(__dirname, 'packages') // Include all CSS from the local package
+          resolveMonorepoRoot('node_modules/@momentum-ui/react-collaboration'), // Include specific node module
+          path.resolve(__dirname, 'packages'), // Include all CSS from the local package
         ],
       },
       {
         test: /\.scss$/,
         use: [
-          "style-loader",  // Injects styles into DOM
-          "css-loader",    // Turns CSS into CommonJS
-          "sass-loader"    // Compiles Sass to CSS
+          'style-loader', // Injects styles into DOM
+          'css-loader', // Turns CSS into CommonJS
+          'sass-loader', // Compiles Sass to CSS
         ],
         include: [
-          resolveMonorepoRoot('node_modules/@momentum-design'), // Include specific node module
-          path.resolve(__dirname, 'packages') // Include all CSS from the local package
+          resolveMonorepoRoot('node_modules/@momentum-ui/react-collaboration'), // Include specific node module
+          path.resolve(__dirname, 'packages'), // Include all CSS from the local package
         ],
       },
       {
         test: /\.(woff|woff2|eot|ttf|otf)$/,
-        include: [resolveMonorepoRoot('node_modules/@momentum-design')],
+        include: [resolveMonorepoRoot('node_modules/@momentum-ui/react-collaboration')],
         type: 'asset/resource',
         generator: {
-          filename: 'fonts/[name][ext][query]'
-        }
+          filename: 'fonts/[name][ext][query]',
+        },
       },
       {
         test: /\.(png|jpg|gif|svg)$/,
-        include: [resolveMonorepoRoot('node_modules/@momentum-design')],
+        include: [resolveMonorepoRoot('node_modules/@momentum-ui/react-collaboration')],
+
         type: 'asset/resource',
         generator: {
-          filename: 'images/[name][ext][query]'
-        }
-      }
+          filename: 'images/[name][ext][query]',
+        },
+      },
     ],
   },
 });