+++ packages/contact-center/station-login/src/station-login/station-login.types.ts
@@ -5,11 +5,6 @@ import {ILogger} from '@webex/cc-store';
  */
 export interface IStationLoginProps {
   /**
-   * The name of the station.
-   */
-  name: string;
-
-  /**
    * Webex instance.
    */
   cc: IContactCenter;
@@ -107,7 +102,6 @@ export interface IStationLoginProps {
 
 export type StationLoginPresentationalProps = Pick<
   IStationLoginProps,
-  | 'name'
   | 'teams'
   | 'loginOptions'
   | 'login'
@@ -126,9 +120,6 @@ export type StationLoginPresentationalProps = Pick<
   showMultipleLoginAlert: boolean;
 };
 
-export type UseStationLoginProps = Pick<
-  IStationLoginProps,
-  'cc' | 'onLogin' | 'onLogout' | 'logger' | 'isAgentLoggedIn' | 'deviceType'
->;
+export type UseStationLoginProps = Pick<IStationLoginProps, 'cc' | 'onLogin' | 'onLogout' | 'logger' | 'deviceType'>;
 
 export type StationLoginProps = Pick<IStationLoginProps, 'onLogin' | 'onLogout'>;