+++ packages/contact-center/store/src/store.ts
@@ -10,7 +10,6 @@ import {
   IStore,
   ILogger,
   IWrapupCode,
-  TASK_EVENTS,
 } from './store.types';
 import {ITask} from '@webex/plugin-cc';
 
@@ -37,11 +36,6 @@ class Store implements IStore {
   constructor() {
     makeAutoObservable(this, {
       cc: observable.ref,
-      currentTask: observable, // Make currentTask observable
-      incomingTask: observable,
-      taskList: observable,
-      wrapupRequired: observable,
-      currentState: observable,
     });
   }