+++ packages/contact-center/store/tests/storeEventsWrapper.ts
@@ -36,6 +36,7 @@ jest.mock('../src/store', () => ({
     setWrapupRequired: jest.fn(),
     setCurrentTheme: jest.fn(),
     setIsAgentLoggedIn: jest.fn(),
+    registerCC: jest.fn(),
   }),
 }));
 
@@ -122,6 +123,21 @@ describe('storeEventsWrapper', () => {
       expect(storeWrapper['store'].currentState).toBe('newState');
     });
 
+    it('should call registerCC', () => {
+      const mockRegisterCC = jest.fn();
+      storeWrapper['store'].registerCC = mockRegisterCC;
+
+      storeWrapper.registerCC();
+      expect(mockRegisterCC).toHaveBeenCalled();
+
+      const mockLogger = {log: jest.fn(), warn: jest.fn(), error: jest.fn(), info: jest.fn(), trace: jest.fn()};
+      storeWrapper.registerCC({
+        cc: {},
+        logger: mockLogger,
+      });
+      expect(mockRegisterCC).toHaveBeenCalledWith({cc: {}, logger: mockLogger});
+    });
+
     it('should setLastStateChangeTimestamp', () => {
       expect(storeWrapper.setLastStateChangeTimestamp).toBeInstanceOf(Function);
 
@@ -156,7 +172,7 @@ describe('storeEventsWrapper', () => {
     });
 
     it('should setWrapupCodes', () => {
-      const mockCodes = [{code1: 'code1'}, {code2: 'code2'}];
+      const mockCodes = [{id: 'code1', name: 'code1'}];
       expect(storeWrapper.setWrapupCodes).toBeInstanceOf(Function);
 
       storeWrapper.setWrapupCodes(mockCodes);
@@ -165,7 +181,7 @@ describe('storeEventsWrapper', () => {
   });
 
   describe('storeEventsWrapper', () => {
-    const mockTask: ITask = ({
+    const mockTask: ITask = {
       data: {
         interactionId: 'interaction1',
         interaction: {
@@ -174,7 +190,7 @@ describe('storeEventsWrapper', () => {
       },
       on: jest.fn(),
       off: jest.fn(),
-    } as unknown) as ITask;
+    } as unknown as ITask;
 
     beforeEach(() => {
       jest.clearAllMocks();
@@ -275,7 +291,7 @@ describe('storeEventsWrapper', () => {
   });
 
   describe('storeEventsWrapper events reactions', () => {
-    const mockTask: ITask = ({
+    const mockTask: ITask = {
       data: {
         interactionId: 'interaction1',
         interaction: {
@@ -284,7 +300,7 @@ describe('storeEventsWrapper', () => {
       },
       on: jest.fn(),
       off: jest.fn(),
-    } as unknown) as ITask;
+    } as unknown as ITask;
 
     const options = {someOption: 'value'};
 
@@ -373,7 +389,7 @@ describe('storeEventsWrapper', () => {
       storeWrapper.handleIncomingTask(mockTask);
 
       expect(setIncomingTaskSpy).toHaveBeenCalledWith(mockTask);
-      expect(mockTask.on).not.toHaveBeenCalledWith;
+      expect(mockTask.on).not.toHaveBeenCalledWith();
     });
 
     it('should handle task assignment', () => {
@@ -607,6 +623,31 @@ describe('storeEventsWrapper', () => {
       expect(storeWrapper['cc'].off).toHaveBeenCalledWith(TASK_EVENTS.TASK_HYDRATE, expect.any(Function));
       expect(storeWrapper['cc'].off).toHaveBeenCalledWith(CC_EVENTS.AGENT_STATE_CHANGE, expect.any(Function));
       expect(storeWrapper['cc'].off).toHaveBeenCalledWith(CC_EVENTS.AGENT_MULTI_LOGIN, expect.any(Function));
+    });
+
+    it('should handle task rejection event and call onTaskRejected with the provided reason', () => {
+      const rejectTask: ITask = {
+        data: {interactionId: 'rejectTest', interaction: {state: 'connected'}},
+        on: jest.fn(),
+        off: jest.fn(),
+      } as unknown as ITask;
+
+      const onTaskRejectedMock = jest.fn();
+      storeWrapper.setTaskRejected(onTaskRejectedMock);
+      const removeSpy = jest.spyOn(storeWrapper, 'handleTaskRemove');
+      storeWrapper['store'].taskList = [];
+
+      storeWrapper.handleIncomingTask(rejectTask);
+      const taskRejectCall = rejectTask.on.mock.calls.find((call) => call[0] === TASK_EVENTS.TASK_REJECT);
+      expect(taskRejectCall).toBeDefined();
+      const rejectCallback = taskRejectCall[1];
+
+      // Simulate rejection event with a specified reason
+      const reason = 'Task Rejected Reason';
+      rejectCallback({reason});
+
+      expect(onTaskRejectedMock).toHaveBeenCalledWith({reason: reason});
+      expect(removeSpy).toHaveBeenCalledWith('rejectTest');
     });
   });
 });